[YS-408] refactor: FooterButtonContainer 컴포넌트 분리
