+++ src/app/home/components/ExperimentPostContainer/FilterContainer/AreaFilter/AreaFilter.css.ts
@@ -180,37 +180,6 @@ export const placeholderArea = style({
   height: '100%',
 });
 
-// Footer 영역들
-export const footerContainer = style({
-  display: 'flex',
-  justifyContent: 'space-between',
-});
-
-export const footerButtonContainer = style({
-  display: 'flex',
-  gap: '0.8rem',
-  alignItems: 'center',
-  justifyContent: 'space-between',
-  width: '100%',
-});
-
-export const infoTextContainer = style({
-  display: 'flex',
-  gap: '0.4rem',
-  alignItems: 'center',
-});
-
-export const infoText = style({
-  ...fonts.label.medium.R13,
-  color: colors.text03,
-});
-
-export const buttonContainer = style({
-  display: 'flex',
-  gap: '0.8rem',
-  alignItems: 'center',
-});
-
 // 버튼 스타일을 recipe로 재사용 (중복되는 padding, borderRadius 등)
 export const buttonRecipe = recipe({
   base: {
