+++ src/app/home/components/ExperimentPostContainer/FilterContainer/AreaFilter/components/FooterButtonContainer/FooterButtonContainer.css.ts
@@ -1,0 +1,35 @@
+import { style } from '@vanilla-extract/css';
+
+import { colors } from '@/styles/colors';
+import { fonts } from '@/styles/fonts.css';
+
+// Footer 영역들
+export const footerContainer = style({
+  display: 'flex',
+  justifyContent: 'space-between',
+});
+
+export const footerButtonContainer = style({
+  display: 'flex',
+  gap: '0.8rem',
+  alignItems: 'center',
+  justifyContent: 'space-between',
+  width: '100%',
+});
+
+export const infoTextContainer = style({
+  display: 'flex',
+  gap: '0.4rem',
+  alignItems: 'center',
+});
+
+export const infoText = style({
+  ...fonts.label.medium.R13,
+  color: colors.text03,
+});
+
+export const buttonContainer = style({
+  display: 'flex',
+  gap: '0.8rem',
+  alignItems: 'center',
+});
